+++ protected/views/site/admin.php
@@ -53,7 +53,7 @@
 
                 <li><a class="btn background-btn" title="Add/update formats of files supported by GigaDB" href="/adminFileFormat/admin">File Formats</a></li>
 
-                <li><a class="btn background-btn" title="Manage GigaDB user accounts" href="/user/admin">Users</a></li>
+                <li><a class="btn background-btn" title="Manage GigaDB user accounts" href="/adminUser/admin">Users</a></li>
 
                 <li><a class="btn background-btn" href="/user/newsletter">Newsletter Subscribers</a></li>