+++ protected/views/user/changePassword.php
@@ -1,13 +1,12 @@
-
 <div class="content">
     <div class="container">
-    <?php
+        <?php
         $this->widget('TitleBreadcrumb', [
-          'pageTitle' => Yii::t('app', 'Change Password'),
-          'breadcrumbItems' => [
-            ['label' => 'Home', 'href' => '/'],
-            ['isActive' => true, 'label' => 'Change Password'],
-          ]
+            'pageTitle'       => Yii::t('app', 'Change Password'),
+            'breadcrumbItems' => [
+                ['label' => 'Home', 'href' => '/'],
+                ['isActive' => true, 'label' => 'Change Password'],
+            ]
         ]);
         ?>
     </div>
@@ -18,67 +17,44 @@
         <div class="col-xs-12 col-md-8 col-md-offset-2">
             <div class="form well user-profile-box">
                 <div class="panel-body p-xs-0 p-sm-0">
-
                     <?php
                     $form = $this->beginWidget('CActiveForm', array(
-                        'id' => 'ChangePassword-form',
+                        'id'                   => 'ChangePassword-form',
                         'enableAjaxValidation' => false,
-                        'htmlOptions' => array('class' => 'form-horizontal'),
+                        'htmlOptions'          => array('class' => 'form-horizontal'),
                     ));
 
-                        echo isset($error) && $error ? '<div class="row">' . $error . '</div>' : '';
+                    echo isset($error) && $error ? '<div class="row">' . $error . '</div>' : '';
 
-                        $this->widget('application.components.controls.PasswordField', [
-                          'form' => $form,
-                          'model' => $model,
-                          'attributeName' => 'password',
-                          'inputOptions' => [
+                    $this->widget('application.components.controls.PasswordField', [
+                        'form'                => $form,
+                        'model'               => $model,
+                        'attributeName'       => 'password',
+                        'inputOptions'        => [
                             'required' => true,
-                          ],
-                          'labelOptions' => ['class' => 'col-xs-12 col-md-4'],
-                          'inputWrapperOptions' => 'col-xs-12 col-md-8'
-                        ]);
-                        $this->widget('application.components.controls.PasswordField', [
-                          'form' => $form,
-                          'model' => $model,
-                          'attributeName' => 'confirmPassword',
-                          'inputOptions' => [
+                        ],
+                        'labelOptions'        => ['class' => 'col-xs-12 col-md-4'],
+                        'inputWrapperOptions' => 'col-xs-12 col-md-8'
+                    ]);
+                    $this->widget('application.components.controls.PasswordField', [
+                        'form'                => $form,
+                        'model'               => $model,
+                        'attributeName'       => 'confirmPassword',
+                        'inputOptions'        => [
                             'required' => true,
-                          ],
-                          'labelOptions' => ['class' => 'col-xs-12 col-md-4'],
-                          'inputWrapperOptions' => 'col-xs-12 col-md-8'
-                        ]);
+                        ],
+                        'labelOptions'        => ['class' => 'col-xs-12 col-md-4'],
+                        'inputWrapperOptions' => 'col-xs-12 col-md-8'
+                    ]);
 
-                        ?>
+                    ?>
 
-                        <div class="form-group checkbox-horizontal <?= $model->hasErrors('newsletter') ? 'has-error' : '' ?>">
-                            <label class="col-xs-12 col-md-4 control-label" for="ChangePasswordForm_newsletter"><?= Yii::t('app', 'Mailing list') ?></label>
-                            <div class="col-xs-12 col-md-8 input-wrapper">
-                                <?php echo $form->checkbox($model, 'newsletter', array('id' => 'ChangePasswordForm_newsletter', 'aria-describedby' => 'newsletterHint')); ?>
-                                </div>
-                            <p id="newsletterHint" class="col-md-8 help-block checkbox-desc">Please tick here to join the GigaDB mailing list to receive news, updates and quarterly newsletters about GigaDB</p>
-                        </div>
+                    <div class="btns-row pull-right">
+                        <a href="/user/view_profile" class="btn background-btn-outline"><?= Yii::t('app', 'Cancel') ?></a>
+                        <?php echo CHtml::submitButton(Yii::t('app', 'Save'), array('class' => 'btn background-btn m-0')); ?>
+                    </div>
 
-                        <div class="form-group checkbox-horizontal <?= $model->hasErrors('terms') ? 'has-error' : '' ?>">
-                            <?= $form->label($model, 'terms', array('class' => 'col-xs-12 col-md-4 control-label')) ?>
-                            <div class="col-xs-12 col-md-8 input-wrapper">
-                                <?php echo $form->checkbox($model, 'terms', array('required' => true, 'aria-required' => 'true', 'aria-describedby' => 'termsHint' . ($model->hasErrors('terms') ? ' termsError' : ''))); ?>
-                            </div>
-                            <div id="termsError" role="alert">
-                                <?php echo $form->error($model, 'terms', array('class' => 'control-error col-md-8 help-block')); ?>
-                            </div>
-                            <p class="col-md-8 help-block checkbox-desc" id="termsHint">Please tick here to confirm you have read and understood our <a href="/site/term#policies">Terms of use</a> and <a href="/site/term#privacy">Privacy Policy</a></p>
-                        </div>
-
-                        <hr class="divider" />
-
-
-                        <div class="btns-row pull-right">
-                            <a href="/user/view_profile" class="btn background-btn-outline"><?= Yii::t('app', 'Cancel') ?></a>
-                            <?php echo CHtml::submitButton(Yii::t('app', 'Save'), array('class' => 'btn background-btn m-0')); ?>
-                        </div>
-
-                  <?php $this->endWidget(); ?>
+                    <?php $this->endWidget(); ?>
                 </div>
             </div>
         </div>